% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{calc_el_hat}
\alias{calc_el_hat}
\title{Calculate el_hat in EM for multivariate LMM}
\usage{
calc_el_hat(yl, B_hat, xl, gl_hat)
}
\arguments{
\item{yl}{a d by 1 matrix of phenotype values for the lth subject}

\item{B_hat}{a matrix of allelotype effect sizes with dimensions a by d}

\item{xl}{a matrix of allelotypes with dimensions d by 1 for a single subject at a single marker}

\item{gl_hat}{a d by 1 matrix of gl_hat values}
}
\description{
Calculate el_hat in EM for multivariate LMM
}
