% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{calc_Sigmabb_hat}
\alias{calc_Sigmabb_hat}
\title{Calculate Sigmabb_hat for use in EM for multivariate LMM}
\usage{
calc_Sigmabb_hat(Xmat, Ve, Vg, Dmat)
}
\arguments{
\item{Xmat}{a genotypes matrix for a single marker, dimension a by n, where a is the number of allelotypes and n the number of subjects}

\item{Ve}{current value of error covariance matrix, with dimension d by d}

\item{Vg}{current value of genetic covariance matrix, with dimension d by d}

\item{Dmat}{a diagonal matrix containing the eigenvalues of the kinship matrix, K = UtDU}
}
\description{
Calculate Sigmabb_hat for use in EM for multivariate LMM
}
