% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{convert_b_hat}
\alias{convert_b_hat}
\title{Convert little b hat matrix (dimension (ad) by 1) to big B hat matrix (dimension a by d), where a is the number of allelotypes at each locus (8 in case of DO) and d is number of traits}
\usage{
convert_b_hat(b_hat, n_traits = 2)
}
\arguments{
\item{b_hat}{ad by 1 matrix of stacked allelotype effect sizes}

\item{n_traits}{number of traits examined simultaneously. Default value is 2}
}
\description{
Convert little b hat matrix (dimension (ad) by 1) to big B hat matrix (dimension a by d), where a is the number of allelotypes at each locus (8 in case of DO) and d is number of traits
}
