% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{update_Vg}
\alias{update_Vg}
\title{Update Vg}
\usage{
update_Vg(Xmat, Vg, Ve, Dmat, y)
}
\arguments{
\item{Xmat}{a genotypes matrix for a single marker, dimension a by n, where a is the number of allelotypes and n the number of subjects}

\item{Vg}{current value of genetic covariance matrix, with dimension d by d}

\item{Ve}{current value of error covariance matrix, with dimension d by d}

\item{Dmat}{a diagonal matrix containing the eigenvalues of the kinship matrix, K = UtDU}

\item{y}{a d by n matrix of phenotype values}
}
\description{
Update Vg
}
