% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{calc_gl_hat}
\alias{calc_gl_hat}
\title{Calculate gl_hat in EM for multivariate LMM}
\usage{
calc_gl_hat(yl, B_hat, Vg, Ve, xl, deltal)
}
\arguments{
\item{Vg}{current value of genetic covariance matrix, with dimension d by d}

\item{Ve}{current value of error covariance matrix, with dimension d by d}

\item{Xmat}{a genotypes matrix for a single marker, dimension a by n, where a is the number of allelotypes and n the number of subjects}

\item{Dmat}{a diagonal matrix containing the eigenvalues of the kinship matrix, K = UtDU}

\item{y}{a d by n matrix of phenotype values}

\item{l}{a number from 1 to n to indicate the subject number}
}
\description{
Calculate gl_hat in EM for multivariate LMM
}
