% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{calc_gl_hat}
\alias{calc_gl_hat}
\title{Calculate gl_hat in EM for multivariate LMM}
\usage{
calc_gl_hat(yl, B_hat, Vg, Ve, xl, deltal)
}
\arguments{
\item{yl}{a column, ie, the lth, from the y matrix}

\item{B_hat}{a matrix of coefficient estimates}

\item{Vg}{current value of genetic covariance matrix, with dimension d by d}

\item{Ve}{current value of error covariance matrix, with dimension d by d}

\item{xl}{a column from the x matrix}

\item{deltal}{lth eigenvalue of K matrix}
}
\description{
Calculate gl_hat in EM for multivariate LMM
}
